% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellcluster.R
\name{cellcluster}
\alias{cellcluster}
\title{Cell clustering}
\usage{
cellcluster(satac, genome, clunum = NULL, perplexity = 30,
  datapath = NULL)
}
\arguments{
\item{satac}{GRanges object or list of GRanges object of scATAC-seq reads. The GRanges should be the middle point of the reads with length of 1 base pair. Use 'satacprocess' to preprocess raw reads.}

\item{genome}{Character variable of either "hg19" or "mm10".}

\item{clunum}{Numeric variable giving the number of clusters. If NULL the cluster number will be determined automatically}

\item{perplexity}{Numeric variable specifying perplexity for tSNE. Reduce perplexity when sample size is small.}

\item{datapath}{Character variable of the path to the customized database (eg myfolder/database.rds). The database can be made using 'makedatabase' function. If not null, 'genome' is ignored.}
}
\value{
A list of three components: tsne results, clustering results and aggregated signal for CRE cluster.
}
\description{
Perform Cell Clustering
}
\details{
This function generates averaged signals for CRE clusters and cluster cells.
}
\examples{
cellcluster(list(cell1=GRanges(seqnames="chr1",IRanges(start=1:100+1e6,end=1:100+1e6)),cell2=GRanges(seqnames="chr2",IRanges(start=1:100+1e6,end=1:100+1e6))),genome="hg19")
}
\author{
Zhicheng Ji, Weiqiang Zhou, Hongkai Ji <zji4@zji4.edu>
}
