% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractfeature.R
\name{extractfeature}
\alias{extractfeature}
\title{Extract features for certain genomic regions}
\usage{
extractfeature(res, region, mode = "overlap", folder = NULL)
}
\arguments{
\item{res}{Matrix outputted by SCATE.}

\item{region}{Dataframe of genomic region. First column: chromosome name. Second column: start position. Third column: end position.}

\item{mode}{Either 'overlap' or 'nearest'. If overlap, only bins that overlap with the regions of interest will be shown. If nearest, the nearest bin to each region will be shown. Order will be the same as the input genomic region.}

\item{folder}{A character value specifying the location where the BED files will be stored. If not NULL, the signal and location of each bin will be saved as BED files in 'folder'. Each cluster has its own BED file, and the name of the BED file is the same as the cluster name. These BED files can be uploaded to UCSC genome browser for visualization.}
}
\value{
A subset of the input matrix
}
\description{
Extract features for certain genomic regions
}
\details{
This function takes as input the recontructed matrix generated by SCATE and a list of genomic regions of interest by the user. It outputs a subset of matrix that overlaps or nearest to the given genomic regions of interest.
}
\examples{
\dontrun{
extractfeature(SCATE(GRanges(seqnames="chr1",IRanges(start=seq_len(100)+1e6,end=seq_len(100)+1e8)),clunum=5000,type='reads',genome="mm10",ncores=10),data.frame(seqnames='chr1',start=1e6,end=1000+1e6))
}
}
\author{
Zhicheng Ji, Weiqiang Zhou, Wenpin Hou, Hongkai Ji <zji4@zji4.edu>
}
