% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedatabase.R
\name{makedatabase}
\alias{makedatabase}
\title{Make customized database}
\usage{
makedatabase(datapath, savepath, bamfile = NULL, cre = NULL,
  genome = "hg19")
}
\arguments{
\item{datapath}{path to the data package folder (e.g. myfolder/hg19/). User must first download the data package to use this function. The data package for hg19 and mm10 can be downloaded from http://jilab.biostat.jhsph.edu/projects/scate/hg19.zip or http://jilab.biostat.jhsph.edu/projects/scate/mm10.zip. The compressed file should be unzipped.}

\item{savepath}{path to save the generated database. e.g. myfolder/database.rds.}

\item{bamfile}{location of bulk DNase-seq bamfiles.}

\item{cre}{dataframe of new CRE sites to be added to the database. First column: chromosome name. Second column: start position. Third column: end position.}
}
\value{
genome either 'hg19' or 'mm10'. Default is 'hg19'.
}
\description{
Make customized database with new bulk DNase-seq data
}
\details{
This function makes a new customized database if users have new bulk DNase-seq data and such information can be contribued to the model building of SCATE.
}
\author{
Zhicheng Ji, Weiqiang Zhou, Hongkai Ji <zji4@zji4.edu>
}
