% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satacprocess.R
\name{satacprocess}
\alias{satacprocess}
\title{Preprocess scATAC-seq}
\usage{
satacprocess(input, type = "bam", pairsingle = "paired", plibsizefilter = 1000)
}
\arguments{
\item{input}{Either a character vector of locations to bam files (when type is bam) or list of GRanges object of scATAC-seq reads (when type is gr).}

\item{type}{Either 'bam' or 'gr'. 'bam' if the input is locations to bam files. 'gr' if the input is a list of GRanges object.}

\item{pairsingle}{Either 'paired' or 'single'. Specify whether the bam file is single- or paired-end. Only works when type='bam'.}

\item{libsizefilter}{Numeric variable giving the minimum library size. scATAC-seq samples with library size smaller than this cutoff will be discarded.}
}
\value{
GRanges object of list of GRanges object after preprocessing.
}
\description{
Preprocessing scATAC-seq samples
}
\details{
This function filters out scATAC-seq with low library size and transforms the reads into middle points of the reads.
}
\examples{
satacprocess(list(cell1=GRanges(seqnames="chr1",IRanges(start=1:100+1e6,end=1:100+1e6)),cell2=GRanges(seqnames="chr2",IRanges(start=1:100+1e6,end=1:100+1e6))),type='gr',libsizefilter=10)
}
\author{
Zhicheng Ji, Weiqiang Zhou, Hongkai Ji <zji4@zji4.edu>
}
