% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCATEpipeline.R
\name{SCATEpipeline}
\alias{SCATEpipeline}
\title{SCATE Pipeline}
\usage{
SCATEpipeline(bamfile, genome, ncores = detectCores(), perplexity = 30)
}
\arguments{
\item{bamfile}{Character vector of bam files to be processed}

\item{genome}{Character variable of either "hg19" or "mm10".}

\item{ncores}{Numeric variable of number of cores to use. If NULL, the maximum number of cores is used.}

\item{perplexity}{Numeric variable specifying perplexity of tSNE. Reduce perplexity when sample size is small.}
}
\value{
A list of two elements. First element is a list generated by cellcluster function, and it contains the cell clustering results. Second element is a matrix generated by SCATE function. Each column is the SCATE result for one cell cluster. Columns are ordered by clusters: first column is the result for cell cluster 1, second column is for cluster 2,...
}
\description{
SCATE pipeline of reading in bam, clustering cell, and performing SCATE
}
\details{
This function takes as input a list of bam files. It then read in the bam files, cluster cells, and performs SCATE for each cell cluster
}
\examples{
f <- list.files(paste0(system.file(package="SCATE"),"/extdata/example"),full.names = T)
SCATEpipeline(f,genome="hg19",perplexity=5)
}
\author{
Zhicheng Ji, Weiqiang Zhou, Hongkai Ji <zji4@zji4.edu>
}
