% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedatabase.R
\name{makedatabase}
\alias{makedatabase}
\title{Make customized database}
\usage{
makedatabase(datapath, savepath, bamfile = NULL, cre = NULL,
  genome = "hg19", genomerange = NULL)
}
\arguments{
\item{datapath}{path to the data package folder (e.g. myfolder/hg19/). User must first download the data package to use this function. The data package for hg19 and mm10 can be downloaded from http://jilab.biostat.jhsph.edu/projects/scate/hg19.zip or http://jilab.biostat.jhsph.edu/projects/scate/mm10.zip. The compressed file should be unzipped. If users do not want to use existing data compendium (e.g. to build a database in a new species), datapath should be set NULL, and 'genome' will be ignored.}

\item{savepath}{path to save the generated database. e.g. myfolder/database.rds.}

\item{bamfile}{location of bulk DNase-seq bamfiles.}

\item{cre}{dataframe of new CRE sites to be added to the database. First column: chromosome name. Second column: start position. Third column: end position.}

\item{genome}{Character variable of either "hg19" or "mm10". Default is 'hg19'. Ignored when datapath is NULL.}

\item{genomerange}{Data frame with two columns. First column is the chromosome and second column is the length of the genome. Only useful when datapath is NULL. Example is https://genome.ucsc.edu/goldenpath/help/hg19.chrom.sizes}
}
\description{
Make customized database with new bulk DNase-seq data
}
\details{
This function makes a new customized database if users have new bulk DNase-seq data and such information can be contribued to the model building of SCATE.
}
\author{
Zhicheng Ji, Weiqiang Zhou, Hongkai Ji <zji4@zji4.edu>
}
