% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakcall.R
\name{peakcall}
\alias{peakcall}
\title{Peak calling}
\usage{
peakcall(feature, genome, flank = 1, fdrcut = 1e-05)
}
\arguments{
\item{feature}{Feature generated by SCATE.}

\item{genome}{Character variable of either "hg19" or "mm10". Should be the same as that used in SCATE function.}

\item{flank}{Numeric variable of the number of flanking bins for each bin. For each bin, an averaged signal of itself and the flanking bins will be calculated and compared to a background distribution.}

\item{fdrcut}{Numeric variable of FDR cutoff. Bins passing the FDR cutoff will be peaks.}
}
\value{
A list of three components: GenomicRanges of peak regions, FDR for peaks and peak signals.
}
\description{
Peak calling function
}
\details{
This function performs peak calling for signal generated by SCATE
}
\examples{
peakcall(SCATE(GRanges(seqnames="chr1",IRanges(start=1:100+1e6,end=1:100+1e6)),genome="hg19"))
}
\author{
Zhicheng Ji, Weiqiang Zhou, Hongkai Ji <zji4@zji4.edu>
}
