% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCATE.R
\name{SCATE}
\alias{SCATE}
\title{Perform SCATE}
\usage{
SCATE(satac, genome, ncores = detectCores())
}
\arguments{
\item{satac}{GRanges object or list of GRanges object of scATAC-seq reads. The GRanges should be the middle point of the reads with length of 1 base pair. Use 'satacprocess' to preprocess raw reads.}

\item{genome}{Character variable of either "hg19" or "mm10".}

\item{ncores}{Numeric variable of number of cores to use. If NULL, the maximum number of cores is used.}
}
\value{
A numeric vector of values generated by SCATE. The length of the vector is the same as the number of bins in the genome.
}
\description{
Single-cell ATAC-seq signal Extration and Enhancement
}
\details{
This function takes as input the scATAC-seq reads and generates enhanced signals.
}
\examples{
SCATE(GRanges(seqnames="chr1",IRanges(start=1:100+1e6,end=1:100+1e6)),genome="hg19")
}
\author{
Zhicheng Ji, Weiqiang Zhou, Hongkai Ji <zji4@zji4.edu>
}
