% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCATE.R
\name{SCATE}
\alias{SCATE}
\title{Perform SCATE}
\usage{
SCATE(satac, genome = "hg19", cluster = NULL, clusterid = NULL,
  datapath = NULL, ncores = detectCores(), verbose = F)
}
\arguments{
\item{satac}{GRanges object or list of GRanges object of scATAC-seq reads. The GRanges should be the middle point of the reads with length of 1 base pair. Use 'satacprocess' to preprocess raw reads.}

\item{genome}{Character variable of either "hg19" or "mm10". Default is 'hg19'.}

\item{cluster}{Numeric vector specifying the cluster of cells. Needs to be named and include all cells in satac. If NULL, SCATE will be run on all cells in satac.}

\item{clusterid}{Numeric number specifying the single cluster to run SCATE. If NULL SCATE will be run on all clusters. Ignored if cluster is NULL. The cluster id must be included in variable 'cluster'.}

\item{datapath}{Character variable of the path to the customized database (eg myfolder/database.rds). The database can be made using 'makedatabase' function. If not null, 'genome' is ignored.}

\item{ncores}{Numeric variable of number of cores to use. If NULL, the maximum number of cores is used.}

\item{verbose}{Either TRUE or FALSE. If TRUE, progress will be displayed.}
}
\value{
A numeric vector or matrix of values generated by SCATE, depending on the number of clusters. The length of the vector or the number of rows of the matrix is the same as the number of bins in the genome. If a matrix, the column names indicate the cluster id.
}
\description{
Single-cell ATAC-seq signal Extration and Enhancement
}
\details{
This function takes as input the scATAC-seq reads and generates enhanced signals. Users can either perform SCATE on clusters of cells or a single group of cells.
}
\examples{
SCATE(GRanges(seqnames="chr1",IRanges(start=1:100+1e6,end=1:100+1e6)),genome="hg19")
}
\author{
Zhicheng Ji, Weiqiang Zhou, Hongkai Ji <zji4@zji4.edu>
}
