% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCATEpipeline.R
\name{SCATEpipeline}
\alias{SCATEpipeline}
\title{SCATE Pipeline}
\usage{
SCATEpipeline(
  bamfile,
  genome = "hg19",
  cellclunum = NULL,
  CREclunum = NULL,
  datapath = NULL,
  ncores = detectCores(),
  perplexity = 30
)
}
\arguments{
\item{bamfile}{Character vector of bam files to be processed}

\item{genome}{Character variable of either "hg19" or "mm10".}

\item{cellclunum}{Numeric variable giving the number of cell clusters when clustering cells. If NULL the cluster number will be determined automatically.}

\item{CREclunum}{Numeric variable giving the number of CRE clusters when running SCATE. If NULL the cluster number will be determined automatically.}

\item{datapath}{Character variable of the path to the customized database (eg myfolder/database.rds). The database can be made using 'makedatabase' function. If not null, 'genome' is ignored.}

\item{ncores}{Numeric variable of number of cores to use. If NULL, the maximum number of cores is used.}

\item{perplexity}{Numeric variable specifying perplexity of tSNE. Reduce perplexity when sample size is small.}
}
\value{
A list of three elements. First element is a list generated by cellcluster function, and it contains the cell clustering results. Second element is a matrix generated by SCATE function. Each column is the SCATE result for one cell cluster. Column names indicate the cluster id. Third element is a list of peaks. Each element is the peak list for one cluster. Name of the element indicates the name of the cluster.
}
\description{
SCATE pipeline of reading in bam, clustering cell, and performing SCATE
}
\details{
This function takes as input a list of bam files. It then read in the bam files, cluster cells, and performs SCATE for each cell cluster
}
\examples{
set.seed(12345)
f <- list.files(paste0(system.file(package="SCATE"),"/extdata/example"),full.names = TRUE)
SCATEpipeline(f,genome="hg19",CREclunum=5000,perplexity=5)
}
\author{
Zhicheng Ji, Weiqiang Zhou, Hongkai Ji <zji4@zji4.edu>
}
